chore: move knowledge files around to match the rest of the project
